Add options to customize analysers directory
Before this commit, only the analysers located in
`{osmose-backend}/analysers` directory are executed.

This commit add an option `analyser-dir`, which allow to change analyser
directory.

Files in this directory are scanned as before, and executed. The
`sys.path` are modified to allow the load of the built-in `analyser`
package. To avoid conflicts, the custom analyser directory (the one
given by the user with `--analyser-dir`) cannot be named `analysers`
(the script stop in this case).

This allow to load the built-in analysers and modules like this:

```
from modules.OsmoseTranslation import T_
from analysers.Analyser_Osmosis import Analyser_Osmosis
```

After the load of analysers, the previous `sys.path` is restored (this
was the behaviour before this commit).